requirements.txt
streamlit
pillow
sqlalchemy
subprocess
pydantic
